Bug fixes and more (see desc.)
Tried to get the font info from the css file used, no matter what file it is; failed. This will get the text shrinking if the name is too big to work perfectly. Right now, you have to manually change the font and font size in the popup.js file every time you change the theme.